refactor: remove the Python plugin
Remove the Python plugin from the extension, allowing a built-in plugin
to take over consuming the Clipboard interface until the Mutter (or
libportal) interface is available.
