+++ .github/workflows/ci.yml
@@ -40,22 +40,7 @@ jobs:
         uses: reviewdog/action-eslint@v1
         with:
           reporter: github-pr-review
-          workdir: src/extension/
-
-  pylint:
-    name: Pylint
-    needs: [pre-test]
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-
-      - name: Lint
-        uses: dciborow/action-pylint@main
-        with:
-          reporter: github-pr-review
-          workdir: src/plugin/gnome-shell/
+          workdir: src/
 
   reuse:
     name: REUSE Compliance
