+++ po/POTFILES
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 # SPDX-FileCopyrightText: Andy Holmes <andrew.g.r.holmes@gmail.com>
 
-src/extension/notification.js
-src/extension/status.js
+src/notification.js
+src/status.js
