+++ po/en.po
@@ -19,44 +19,44 @@ msgstr ""
 "Plural-Forms: nplurals=2; plural=(n != 1);\n"
 
 #. TRANSLATORS: A notification button to show the quick-reply entry
-#: src/extension/notification.js:59
+#: src/notification.js:59
 msgid "Reply"
 msgstr "Reply"
 
 #. TRANSLATORS: A reply entry in a notification
-#: src/extension/notification.js:69
+#: src/notification.js:69
 msgid "Type a message"
 msgstr "Type a message"
 
-#: src/extension/status.js:262 src/extension/status.js:431
+#: src/status.js:262 src/status.js:431
 msgid "Devices"
 msgstr "Devices"
 
-#: src/extension/status.js:287 src/extension/status.js:389
-#: src/extension/status.js:393
+#: src/status.js:287 src/status.js:389
+#: src/status.js:393
 msgid "All Devices"
 msgstr "All Devices"
 
 #. TRANSLATORS: This is the number of connected devices
-#: src/extension/status.js:364
+#: src/status.js:364
 #, javascript-format
 msgid "%d Connected"
 msgid_plural "%d Connected"
 msgstr[0] "%d Connected"
 msgstr[1] "%d Connected"
 
-#: src/extension/status.js:383
+#: src/status.js:383
 msgid "Valent must be installed to connect and sync devices"
 msgstr ""
 
-#: src/extension/status.js:384
+#: src/status.js:384
 msgid "Help"
 msgstr ""
 
-#: src/extension/status.js:388
+#: src/status.js:388
 msgid "No available or connected devices"
 msgstr ""
 
-#: src/extension/status.js:392
+#: src/status.js:392
 msgid "Turn on to connect to devices"
 msgstr ""
