+++ src/clipboard.js
