--- src/plugin/gnome-shell.plugin.in
@@ -1,17 +1,0 @@
-# SPDX-License-Identifier: GPL-3.0-or-later
-# SPDX-FileCopyrightText: Andy Holmes <andrew.g.r.holmes@gmail.com>
-
-[Plugin]
-Module=gnome-shell
-Name=GNOME Shell
-Description=Integration with GNOME Shell
-Icon=application-x-addon-symbolic
-Builtin=false
-Version=@PLUGIN_VERSION@
-Website=https://github.com/andyholmes/gnome-shell-extension-valent
-Help=https://github.com/andyholmes/gnome-shell-extension-valent/wiki
-Hidden=false
-Loader=python3
-X-ClipboardAdapterPriority=50
-X-SessionAdapterPriority=50
-
