+++ tests/meson.build
@@ -12,14 +12,3 @@ if eslint.found()
   )
 endif
 
-# Lint Python
-pylint = find_program('pylint', required: false)
-
-if pylint.found()
-  test('Pylint', pylint,
-           args: join_paths(meson.project_source_root(), 'src', 'plugin', 'gnome-shell'),
-    is_parallel: true,
-          suite: 'lint',
-  )
-endif
-
